Added a perlin noise function
Perlin Noise function determines the angle for the frost particle instead of hard coding it. This gives you different versions on each refresh instead of the same one over and over again. Next is to animate the growth of the frost across the screen using another function that uses Perlin Noise.